fix module local
